Refactor Command class to have child class TasksCmd
Currently Command class has all commands.

However some commands, particularly the tasks commands have
specialised tasks such as parsing inputs from the data file.

Let's,
* Extend the Command class with a TasksCmd class.
* Refactro all task related commands.
