+++ src/main/java/botling/commands/commandtypes/ByeCmd.java
@@ -5,8 +5,6 @@ import botling.commands.CmdConst;
 import botling.commands.CommandColor;
 import botling.messagegenerator.MsgGen;
 
-import java.io.BufferedReader;
-
 /**
  * Parses bye commands.
  */
@@ -21,12 +19,5 @@ public class ByeCmd implements Command {
             return MsgGen.bye(cmdColor);
         }
         return MsgGen.unknownCmd(cmdColor);
-    }
-
-    /**
-     * Dummy method.
-     */
-    public void restore(BufferedReader reader, TaskList tasks) {
-        // Do nothing.
     }
 }
