+++ src/main/java/botling/commands/commandtypes/DeleteCmd.java
@@ -7,8 +7,6 @@ import botling.commands.CommandColor;
 import botling.commands.ValConstants;
 import botling.messagegenerator.MsgGen;
 
-import java.io.BufferedReader;
-
 /**
  * Parses delete commands.
  */
@@ -40,12 +38,5 @@ public class DeleteCmd implements Command {
         return MsgGen.unknownSyntax(CmdConst.CMD_DELETE.getString(),
                 CmdConst.MSG_INVALID_CMD_MARK.getString()
                         + String.valueOf(tasks.size()), cmdColor);
-    }
-
-    /**
-     * Dummy method.
-     */
-    public void restore(BufferedReader reader, TaskList tasks) {
-        // Do nothing.
     }
 }
