+++ src/main/java/botling/commands/commandtypes/ListCmd.java
@@ -5,8 +5,6 @@ import botling.commands.CmdConst;
 import botling.commands.CommandColor;
 import botling.messagegenerator.MsgGen;
 
-import java.io.BufferedReader;
-
 /**
  * Parses list commands.
  */
@@ -20,12 +18,5 @@ public class ListCmd implements Command {
             return MsgGen.list(tasks.list(), cmdColor);
         }
         return MsgGen.unknownCmd(cmdColor);
-    }
-
-    /**
-     * Dummy method.
-     */
-    public void restore(BufferedReader reader, TaskList tasks) {
-        // Do nothing.
     }
 }
