+++ src/main/java/botling/gui/SelectableTextFlow.java
@@ -2,7 +2,6 @@ package botling.gui;
 
 import java.util.List;
 
-import javafx.geometry.Insets;
 import javafx.scene.Node;
 import javafx.scene.control.ScrollBar;
 import javafx.scene.control.TextArea;
@@ -56,7 +55,7 @@ public class SelectableTextFlow extends TextFlow {
         textArea.setLayoutX(6); // hardcoded
         textArea.setLayoutY(6); // hardcoded
         textArea.resize(getWidth() - 11.8, getHeight()); // hardcoded
-        ScrollBar scrollBarv = (ScrollBar)textArea.lookup(".scroll-bar:vertical");
+        ScrollBar scrollBarv = (ScrollBar) textArea.lookup(".scroll-bar:vertical");
         scrollBarv.setDisable(true);
     }
 }
